+++ rim-application/src/lib.rs
@@ -2,11 +2,5 @@ pub mod action;
 pub mod action_handler;
 pub mod command;
 pub mod config;
-pub mod display_geometry {
-	pub use rim_domain::display_geometry::*;
-}
 pub mod ports;
-pub mod preview {
-	pub use rim_domain::preview::*;
-}
 pub mod state;